feat: initialize nix flake for project
Initialize Nix Flake configuration for project using `nix flake init`.

This declaratively defines the development environment for this project
by explicitly mentioning the requires dependencies to run the program
(in the 'shell.nix' file) as well as locking those dependencies to a
consistent version across devices (using the 'flake.lock' file).

Additionally, automatically trigger the enabling of this developer
environment using nix-direnv to trigger `nix develop` upon entering the
project directory. This is done through the '.envrc' file and requires
that nix-direnv is installed on the system.