feat: add data to database
Add required data to database. Data was scraped from Bulbapedia using
the Python script in 'scripts/scrapeRecipeData.py'.

The SQL tables were structured with a many-to-many relationship with an
intermediate table to join the dishes and ingredients. There are
currently a total of 16 rows in the 'Ingredient' table, 60 rows in the
'Dish' table and 960 rows in the 'DishIngredient' table.

A sample DB query was added to 'src/pages/index.astro' to demonstrate
how data will be fetched from the database.